+++ src/modules/matrix.js
@@ -1,6 +1,7 @@
+import { checktypes } from 'es7-checktypes-decorator';
+
 import { WrongSchemaError } from '../errors.js';
 import { arrayEqual, iter } from '../reusables.js';
-import { checktypes } from '../checkTypes.js';
 
 /**
 * Matrix module for DataFrame, providing basic mathematical matrix computations.